+++ CHANGELOG.md
@@ -45,6 +45,8 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 - :tada: **Enhancement** added **`StandardErrorCriterion`**
 - :tada: **Enhancement** added **`VarianceCriterion`**
 - :tada: **Enhancement** added **`AverageCriterion`**
+- :tada: **Enhancement** added **`TVBIndicator`**
+- :tada: **Enhancement** added **`REXIndicator`** with SMA
 
 ## 0.14 (released April 25, 2021)